+++ examples/async_example.py
@@ -43,6 +43,7 @@ custom_vocabularies = [mit_professors, other_scientists]
 #                                    custom_vocabularies=None,
 #                                    filter_profanity=False,
 #                                    remove_disfluencies=False,
+#                                    language=None,
 #                                    delete_after_seconds=None)
 
 
@@ -56,6 +57,7 @@ job = client.submit_job_url(url,
                             custom_vocabularies=custom_vocabularies,
                             filter_profanity=False,
                             remove_disfluencies=False,
+                            language=None,
                             delete_after_seconds=None)
 
 print("Submitted Job")