+++ src/rev_ai/apiclient.py
@@ -46,6 +46,7 @@ class RevAiAPIClient(BaseClient):
             custom_vocabularies=None,
             filter_profanity=False,
             remove_disfluencies=False,
+            language=None,
             delete_after_seconds=None):
         """Submit media given a URL for transcription.
         The audio data is downloaded from the URL.
@@ -70,6 +71,8 @@ class RevAiAPIClient(BaseClient):
             itself toward.
         :param filter_profanity: whether to mask profane words
         :param remove_disfluencies: whether to exclude filler words like "uh"
+        :param language: specify language using ISO 639-1 2-letter language
+                         code
         :param delete_after_seconds: number of seconds after job completion when job is auto-deleted
         :returns: raw response data
         :raises: HTTPError
@@ -81,7 +84,8 @@ class RevAiAPIClient(BaseClient):
                                                    callback_url, skip_diarization,
                                                    skip_punctuation, speaker_channels_count,
                                                    custom_vocabularies, filter_profanity,
-                                                   remove_disfluencies, delete_after_seconds)
+                                                   remove_disfluencies, language,
+                                                   delete_after_seconds)
 
         response = self._make_http_request(
             "POST",
@@ -101,6 +105,7 @@ class RevAiAPIClient(BaseClient):
             custom_vocabularies=None,
             filter_profanity=False,
             remove_disfluencies=False,
+            language=None,
             delete_after_seconds=None):
         """Submit a local file for transcription.
         Note that the content type is inferred if not provided.
@@ -124,6 +129,8 @@ class RevAiAPIClient(BaseClient):
             a phrase you would like the speech recognition to bias itself toward.
         :param filter_profanity: whether to mask profane words
         :param remove_disfluencies: whether to exclude filler words like "uh"
+        :param language: specify language using ISO 639-1 2-letter language
+                         code
         :param delete_after_seconds: number of seconds after job completion when job is auto-deleted
         :returns: raw response data
         :raises: HTTPError
@@ -134,7 +141,8 @@ class RevAiAPIClient(BaseClient):
         payload = self._create_job_options_payload(None, metadata, callback_url, skip_diarization,
                                                    skip_punctuation, speaker_channels_count,
                                                    custom_vocabularies, filter_profanity,
-                                                   remove_disfluencies, delete_after_seconds)
+                                                   remove_disfluencies, language,
+                                                   delete_after_seconds)
 
         with open(filename, 'rb') as f:
             files = {
@@ -373,6 +381,7 @@ class RevAiAPIClient(BaseClient):
             custom_vocabularies=None,
             filter_profanity=None,
             remove_disfluencies=None,
+            language=None,
             delete_after_seconds=None):
         payload = {}
         if media_url:
@@ -394,6 +403,8 @@ class RevAiAPIClient(BaseClient):
             payload['filter_profanity'] = filter_profanity
         if remove_disfluencies:
             payload['remove_disfluencies'] = remove_disfluencies
+        if language:
+            payload['language'] = language
         if delete_after_seconds is not None:
             payload['delete_after_seconds'] = delete_after_seconds
         return payload